Module tetrs::block

Defines a single tetrimino and actions for modifying it.

A Block consists of a type, location and rotation. Each block does not store an internal reference to a Field, but it is required in most useful functions.

use tetrs::import::*;

let field = Field::new();
// A block must be spawned on a field
let block1 = Block::new(block::Id::I, &field);

// This position can be overridden if required (still requires a field)
let block2 = Block::with_options(block::Id::I, &field, BlockOptions {
                x: Some(3), y: Some(4),
                ..Default::default()
            });

Structs

Block
A struct representing a single tetrimino.

BlockOptions
Optional values which can be set when initializing a Block.

Enums

Direction
A movement along one of the four directional axes.

Id
The identifier for a particular Block.

Rotation
Represents all rotation statuses a block can be. This is used both as a rotation state, and to indicate how much relative movement shoud be applied for various functions.