Struct tetrs::block::Block

pub struct Block {
    pub x: i32,
    pub y: i32,
    pub id: Type,
    pub r: Rotation,
    pub data: &'static [(usize, usize)],
}

A block.

All coordinates are taken from the top-left of the field, increasing y moves the block down, increasing x moves the block right.

This is done to preserve compatibility with how most GUI systems provide their coordinates.

Fields

x

X-coordinate of the piece

y

Y-coordinate of the piece

id

Type of the block

r

Rotation state of the block

data

Block offset data

Methods

impl Block

fn new(id: Type, r: Rotation) -> Block

Construct a new block from the specified input parameters.

Examples

use tetrs::block::Block;

let block = Block::new(tetrs::block::Type::I, tetrs::Rotation::R90);

fn new_at(id: Type, r: Rotation, (x, y): (usize, usize)) -> Block

Construct a block at the specific location on the field.

If the location is invalid and is filled, then return None. This is not often used, but can be useful for specific cases.

fn reset(self) -> Block

Consumes the current block returning a new block that has been reset (respawned).

This is useful when implementing a hold function

fn leading(&self) -> (i32, i32)

Returns a tuple with the minimum x, y values of the current block.

fn trailing(&self) -> (i32, i32)

Returns a tuple with the maximum x, y values of the current block

fn collision(&self, field: &Field) -> bool

Checks if the current block collides with anything on field.

fn shift(&mut self, field: &Field, direction: Direction) -> bool

Shift the block one place by the specified direction.

Examples

use tetrs::field::Field;
use tetrs::block::Block;

let field = Field::new();
let mut block = Block::new(tetrs::block::Type::Z, tetrs::Rotation::R0);
block.shift(&field, tetrs::Direction::Left);

fn shift_extend(&mut self, field: &Field, direction: Direction)

Repeatedly shift a block as far as we can until a collision occurs.

Examples

use tetrs::field::Field;
use tetrs::block::Block;

let field = Field::new();
let mut block = Block::new(tetrs::block::Type::Z, tetrs::Rotation::R0);
block.shift_extend(&field, tetrs::Direction::Left);

fn rotate_with_offset(&mut self, field: &Field, rotation: Rotation, (x, y): (i32, i32)) -> bool

Rotate the block by the specified amount and apply an offset. This is a helper function for calculating wallkick values easily.

fn rotate(&mut self, field: &Field, rotation: Rotation) -> bool

Rotate the block by the specified amount.

Examples

use tetrs::field::Field;
use tetrs::block::Block;

let field = Field::new();
let mut block = Block::new(tetrs::block::Type::Z, tetrs::Rotation::R0);
block.rotate(&field, tetrs::Rotation::R90);

fn at(&self, (a, b): (usize, usize)) -> bool

Check if the block occupies the specified offset.

fn ghost(&self, field: &Field) -> Block

Return a new block that is this blocks ghost.

Examples

use tetrs::field::Field;
use tetrs::block::Block;

let field = Field::new();
let block = Block::new(tetrs::block::Type::Z, tetrs::Rotation::R0);
let ghost = block.ghost(&field);

fn data(id: Type, r: Rotation) -> &'static [(usize, usize)]

Return the block data for the specified type.

fn offset(id: Type, r: Rotation) -> (usize, usize)

Return the block offsets for the specified type.

fn offset_to_first(id: Type, r: Rotation) -> (usize, usize)

Return the offset to the first piece. This is slightly different from offset which itself returns the offsets of empty rows/cols.