Struct tetrs::block::BlockOptions

pub struct BlockOptions {
    pub x: Option<i32>,
    pub y: Option<i32>,
    pub rotation: Rotation,
    pub rotation_system: &'static RotationSystem,
}

Optional values which can be set when initializing a Block.

The default values are:

BlockOptions {
    x: None,
    y: None,
    rotation: Rotation::R0,
    rotation_system: rotation_system::SRS
}

If an x or y value is Some(..) then it will override the fields spawn position, which is usually used on Block construction.

Examples

use tetrs::import::*;

// Has x: None, y: None, rotation: Rotation::R0, rotation_system: "dtet"
let options = BlockOptions {
    rotation_system: rotation_system::new("dtet").unwrap(),
    ..Default::default()
};

Fields

x

y

rotation

rotation_system