Trait tetrs::block::BlockBuilder

pub trait BlockBuilder {
    fn set_position(self, position: (i32, i32)) -> Block;
    fn set_rotation(self, rotation: Rotation) -> Block;
    fn set_field(self, field: &Field) -> Block;
    fn set_rotation_system<R: RotationSystem>(self, rs: &'static R) -> Block;
}

Traits for building a block.

This is used to override given default block values to more specific values if required.

These should only be used on construction of a new block with block::new(). This is not enforced, but limited checking is done internally and the results may not be as one would expect.

No collision testing is done on the resulting block, so it is up to the caller to perform this.

Examples

use tetrs::block::{Type, Rotation, Block, BlockBuilder};

let block = Block::new(Type::I)
                  .set_rotation(Rotation::R270);
                  .set_position((5, 10))
                  .set_rotation_system(rotation::SRS::new());

See block::new() for what default values are used.

Required Methods

fn set_position(self, position: (i32, i32)) -> Block

Alter the initial position of the block.

fn set_rotation(self, rotation: Rotation) -> Block

Alter the initial rotation of the block.

fn set_field(self, field: &Field) -> Block

Alter the initial position of the block, setting it to the spawn position as specified by field.

fn set_rotation_system<R: RotationSystem>(self, rs: &'static R) -> Block

Alter the default RotationSystem used by the block.