Module tetrs::controller

An abstract controller for specifying actions.

This controller acts like a bridge between any physical input and the input which the internal gamestate can understand.

When a key is pressed, this can be translated to a call to the controllers activate function, and when the key is removed, this corresponds to a call to the deactivate function.

Structs

Controller
A controller stores the internal state as a series of known actions.

Enums

Action
Actions which are understood by the controller.

Type Definitions

CAarray
'Controller Active' array

CTarray
'Controller Time' array.