Struct tetrs::controller::Controller

pub struct Controller {
    pub time: [usize; 8],
    pub active: [bool; 8],
}

A controller stores the internal state as a series of known actions.

The active status of each action is stored, along with how long each action has been active for. At its simplest, this/ controller parallels the keystate of a keyboard.

Fields

time

The length each action has occured for in ticks.

active

Which actions are currently active.

Methods

impl Controller

fn new() -> Controller

Return a new controller instance.

The controller specified will start with all actions in the inactive state.

Examples

use tetrs::controller;

let mut controller = controller::Controller::new();

fn active(&self, action: Action) -> bool

Query if an action is currently active.

fn time(&self, action: Action) -> usize

Query how long an action has been active for.

fn activate(&mut self, action: Action)

Activate the specified action.

The action will be set to active and all subsequent updates will now trigger this actions timer. Reactivating an already active action has no effect and will not restart the timer.

Examples

use tetrs::controller;

let mut controller = controller::Controller::new();
controller.activate(controller::Action::MoveLeft);

fn deactivate(&mut self, action: Action)

Deactivate the specified action.

The action will be set to inactive. Deactivating an already inactive action has no effect.

Examples

use tetrs::controller;

let mut controller = controller::Controller::new();
controller.deactivate(controller::Action::MoveLeft);

fn deactivate_all(&mut self)

Deactivate all actions.

This is useful when calculating actions based on the state of an object, rather than explicitly via events.

This does not reset the internal time of each action to 0.

Examples

use tetrs::controller::{Controller, Action};

let mut controller = Controller::new();
controller.deactivate_all();

fn update(&mut self)

Update all active actions and increment their timers.

Examples

use tetrs::controller::{Action, Controller};

let mut controller = Controller::new();
controller.activate(Action::MoveRight);
// active[Action::MoveRight] == 0
controller.update();
// active[Action::MoveRight] == 1