Struct tetrs::engine::Engine

pub struct Engine {
    pub co: Controller,
    pub rd: Box<Randomizer>,
    pub wk: &'static Wallkick,
    pub rs: &'static RotationSystem,
    pub fd: Field,
    pub bk: Block,
    pub hd: Option<Id>,
    pub op: EngineSettings,
    pub st: Statistics,
    pub hs: History,
    pub running: bool,
    pub mspt: u64,
    pub tick_count: u64,
    // some fields omitted
}

Stores the internal engine details.

This is largely segmented into components EngineSettings, EngineInternal to reduce the overall complexity and namespace various features.

Most names are condensed a lot to provide shorter references.

Fields

co

Controller which is used by the engine

rd

The randomizer being used.

wk

The wallkick object being used.

rs

The rotation system used by this engine.

fd

The field which the game is played on

bk

The active block

hd

The current hold block (this doesn't store an actual block right now)

op

Settings used internally by the engine

st

Statistics of the current game

hs

The input history of the game

running

Is the game running

mspt

How many milliseconds occur per game tick.

tick_count

How many ticks have elapsed this game

Methods

impl Engine

fn update(&mut self)

The main update phase of the engine.

This handles DAS and all other internal complications based on the current controller state only.

Each call to update is expected to take place in ~mspt ms. It is up to the caller to manage the update lengths appropriately.

fn new(options: EngineOptions) -> Engine

Construct a new Engine from an EngineOptions instance.