Struct tetrs::field::Field

pub struct Field {
    pub width: usize,
    pub height: usize,
    pub hidden: usize,
    pub spawn: (i32, i32),
    pub data: Vec<Vec<Id>>,
}

A Field is simply a 2-D Vec with some corresponding options.

Internally we use a Vec of Vec's for easier resizing, but all rows must be of equal length, so it can be thought of as rectangular for all purposes.

Note

If one wants to only iterate over the viewable portion of the array, this can be done with the following bounds:

for y in self.hidden..self.height {
    ...
}

instead of the more presumed

for y in 0..self.height {
}

This is because the height field includes the specified hidden portion.

Fields

width

The width of the field.

height

The height of the field.

hidden

The height of the hidden region of the field.

This is a subregion of height, and does not actually add any more height to the field.

spawn

The initial spawn of a Block on this field.

data

The current field state.

Methods

impl Field

fn new() -> Field

Construct a Field object with default values.

fn with_options(options: FieldOptions) -> Field

Construct a Field object with specific values.

Examples

use tetrs::import::*;

let field = Field::with_options(FieldOptions {
                width: 15, height: 22,
                ..Default::default()
            });

fn clear_lines(&mut self) -> usize

Clear lines from the field and return the number cleared.

fn freeze(&mut self, block: Block)

Freeze a block into place on the field. This takes ownership of the block to ensure it cannot be used again.

Examples

use tetrs::import::*;

let mut field = Field::new();
let mut block = Block::new(block::Id::I, &field);

field.freeze(block);

// block.shift(Direction::Right); // Compile Error

fn get(&self, (x, y): (usize, usize)) -> Id

Return the value at the specified field location.

This currently is a usize value which corresponds to some Id of Block.

Examples

use tetrs::import::*;

let field = Field::new();
let value = field.get((5, 10));

fn occupies(&self, (x, y): (usize, usize)) -> bool

Return true if the value at the specified location is non-empty.

This is a convenience function which queries at and checks if the result is of empty Id.

If the input values are outside the range of the field then true is returned. Explicit bounds checking is left up the user.