Struct tetrs::history::History

pub struct History {
    // some fields omitted
}

Manages the history state of a particular game.

This at its simplest is a sequence of (u64, Action) tuples which represent particular the entire game input state.

This can be used to recreate games or provide other interesting statistics. This at some point could be combined with the statistics class, where statistic could be some Event-like structure which would allow time-based statistics tracking.

Methods

impl History

fn new() -> History

Construct a new empty history sequence

fn update(&mut self, controller: &Controller)

Update the history state with a controller snapshot.

fn get_sequence(&self) -> &[Event]

Return the current event sequence