Struct tetrs::options::Options

pub struct Options {
    pub das: u64,
    pub are: u64,
    pub gravity: u64,
    pub arr: u64,
    pub has_hold: bool,
    pub hold_limit: u64,
    pub has_hard_drop: bool,
    pub has_soft_drop: bool,
    pub soft_drop_speed: u64,
    pub preview_count: u64,
}

Stores a number of internal options that may be useful during a games execution.

Unlike NullpoMino these are never tied to a name. This can be handled by the caller if required.

Currently Options do not manage the randomizer/rotation system etc. Need to determine exactly what is required.

Fields

das

DAS setting (in ms)

are

ARE time (in ms)

gravity

Gravity (in ms). How many ms must pass for block to fall

arr

Auto-repeat-rate (in ms)

has_hold

Is hold enabled

hold_limit

How many times can we hold

has_hard_drop

Is hard drop allowed

has_soft_drop

Has soft drop

soft_drop_speed

The speed soft drop works

preview_count

Maximum number of preview pieces

Methods

impl Options

fn new() -> Options

Construct a new Options value.

If more specific options are required, make use of the Default trait impl.

use options::Options;

// Standard initialization
let options = Options::new();

// Specific initialization overriding only the `das` field
let options = Options {
    das: 110,
    ..Default::default()
}