Module tetrs::randomizer

Implements a randomizer.

Randomizers currently only are required to implement the Iterator trait. Preview pieces are managed by the caller, although this could be something added as a trait.

It is arguable whether Randomizer should imply Iterator. A Randomizer is often only processed a single element at a time, and cannot be used in an iterator fashion due to borrow rules.

Also, all Randomizer's should return infinite sequences so we can remove the required unwrap on manual calls to next.

Structs

BagRandomizer
A generic bag randomizer.

GameboyRandomizer
A generic memoryless randomizer. This generates a completely arbitrary sequence of Type's.

MemorylessRandomizer
A generic memoryless randomizer. This generates a completely arbitrary sequence of Type's.

TGM1Randomizer
A TGM1 randomizer. This generates a completely arbitrary sequence of Type's.

TGM2Randomizer
A TGM2 randomizer. This generates a completely arbitrary sequence of Type's.

Traits

Randomizer
A randomizer must implement an iterator, plus a preview function which returns a number of lookahead pieces.