Trait tetrs::rotation::RotationSystem

pub trait RotationSystem {
    fn data(&self, ty: Type, rotation: Rotation) -> &'static [(usize, usize)];

    fn min(&self, id: Type, r: Rotation) -> (usize, usize) { ... }
    fn max(&self, id: Type, r: Rotation) -> (usize, usize) { ... }
    fn minp(&self, id: Type, r: Rotation) -> (usize, usize) { ... }
}

The RotationSystem trait is implmented by all rotation systems.

When implementing RotationSystem, the only thing that is required to implement are the offset values for each of the main blocks I, J, L, S, Z, T, O.

See the srs.rs file for an example.

Required Methods

fn data(&self, ty: Type, rotation: Rotation) -> &'static [(usize, usize)]

Returns a static array of offset values for the specified Type and Rotation.

Provided Methods

fn min(&self, id: Type, r: Rotation) -> (usize, usize)

Returns a tuple containing the leading empty (x, y) columns.

Examples

use tetrs::block::{Rotation, Type};
use tetrs::rotation::{self, RotationSystem};

let rs = rotation::SRS{};

// An L-block can have the following representation
// .#.
// .#.
// .##

let (x1, y1) = rs.min(Type::L, Rotation::R90);
assert_eq!((x1, y1), (1, 0));

// An I-block can have the following representation
// ....
// ....
// ####
// ....

let (x2, y2) = rs.min(Type::I, Rotation::R180);
assert_eq!((x2, y2), (0, 2));

fn max(&self, id: Type, r: Rotation) -> (usize, usize)

Returns an (x, y) tuple containing the maximum offsets for the specified block.

Examples

use tetrs::block::{Rotation, Type};
use tetrs::rotation::{self, RotationSystem};

let rs = rotation::SRS{};

// An L-block can have the following representation
// .#.
// .#.
// .##

let (x1, y1) = rs.max(Type::L, Rotation::R90);
assert_eq!((x1, y1), (2, 2));

// An I-block can have the following representation
// ....
// ....
// ####
// ....

let (x2, y2) = rs.max(Type::I, Rotation::R180);
assert_eq!((x2, y2), (3, 2));

fn minp(&self, id: Type, r: Rotation) -> (usize, usize)

Returns the minimum offset of the first piece in a block.

Return the offset from the (x, y) bounding coordinate to the first non-empty piece in a block. This row by row from y = 0 onwards.

use tetrs::block::{Rotation, Type};
use tetrs::rotation;
use tetrs::rotation::RotationSystem;

// The piece marked '@' is the first encountered piece.
// ...
// @##
// .#.

let rs = rotation::SRS{};

let (x1, y1) = rs.minp(Type::T, Rotation::R180);
assert_eq!((x1, y1), (0, 1));