Module tetrs::rotation_system

This modules provides an interface for dealing with different block offsets.

Offsets refer to particular rotation specifications. For example, the SRS and Akira style rotation systems which each contain different offset values which can both be used if they implement the RotationSystem trait.

Reexports

pub use self::srs::SRS;

pub use self::arika::Arika;

pub use self::tengen::Tengen;

pub use self::dtet::DTET;

Modules

arika
Specifies the block values for the Arika rotation system.

dtet
Rotation offsets for the DTET rotation system.

srs
Specifies the block values for the SRS rotation system.

tengen
Specifies the block values for the Tengen rotation system.

Traits

RotationSystem
The RotationSystem trait is implmented by all rotation systems.

Functions

new
Factory function for constructing a rotation system from name.