Module tetrs::schema

Methods for converting to and from a textual field representation.

This is tremendously useful for writing complex tests.

Examples:

use tetrs::schema::Schema;

let (field, mut block) = Schema::from_string("
    |          |
    |  #       |
    | # @@     |
    |## @@     |
    ------------
").to_state();

block.rotate(&field, tetrs::Rotation::R90);

assert_eq!(Schema::from_state(&field, &block), Schema::from_string("
    |          |
    |  #       |
    | # @@     |
    |#  @@     |
    ------------
"));

This can also potentially be useful for specifying initial fields of play. This requires slightly more work in combining with other field initialization parameters.

Structs

Schema
A schema is a simple 2d textual representation of a field and a block. It allows conversion from a string, and also from a (&Field, &Block) and bridges the gap between these two components.