Struct tetrs::schema::Schema

pub struct Schema {
    pub width: usize,
    pub height: usize,
    // some fields omitted
}

A schema is a simple 2d textual representation of a field and a block. It allows conversion from a string, and also from a (&Field, &Block) and bridges the gap between these two components.

This is stored in (y, x) index order to make some operations easier. This is only an internal detail, but care needs to be taken when interfacing with any other components (which are mostly (x, y) indexed).

Fields

width

The current width of the schema

height

The current height of the schema

Methods

impl Schema

fn from_state(field: &Field, block: &Block) -> Schema

Construct a schema representation from an game primitives.

fn from_string(field: &str) -> Schema

Construct a schema representation from an input string.

An input string is considered as line-seperated, with the field lying between pairs of | characters. Leading and trailing whitespace is ignored so different strings may produce the same schema.

Examples

use tetrs::schema::Schema;

let schema1 = Schema::from_string("
    |         |
    |    #    |
    -----------
", rotation_system::new("dtet"));

let schema2 = Schema::from_string("|          |
   |    #     |
 ------------");

 assert_eq!(schema1, schema2); // True

fn to_state(&self, rotation_system: &'static RotationSystem) -> (Field, Block)

Constuct state objects from a given schema. This is slightly finicky and there are a few cases to consider.

It is possible that the schema itself is not well-formed and there is no suitable state to represent it, in which case we panic.

It is up to the caller to ensure input validity. Since this is usually always used with static strings it is straight-forward enough in most cases.

Examples

use tetrs::schema::Schema;
use tetrs::rotation_system;

let rs = rotation_system::new("srs");

let schema1 = Schema::from_string("
    |         |
    |    @    |
    -----------
");

// let (field, block) = schema1.to_state(rs); // Panic!

let schema2 = Schema::from_string("
    |    @    |
    |   @@@   |
    -----------
");
let (field, block) = schema2.to_state(); // Okay

fn to_string(&self) -> String

Construct a visual representation from a given schema.

All output schema have their leading rows truncated for brevity.