Crate tetrs

The tetrs library provides a number of low-level tasks related to movement of blocks. The code aims to be correct and provide easy extension for new input.

Intra-module dependencies are as reduced in scope as possible. For example, a block itself has no knowledge of a wallkick, but provides functionality in terms of rotations with offset to allow for easy use with wallkicks.

A higher-level abstraction in terms of a game-engine is provided. This provides DAS-like behaviour and such.

Modules

block
A block represents a single tetrimino and is always tied to a specific field instance.

controller
An abstract controller for specifying actions.

field
A generic field for use with a game.

randomizer
Implements a randomizer.

schema
Methods for converting to and from a textual field representation.

wallkick
Implements a wallkick test

Enums

Direction
Specifies a particular direction of movement.

Rotation
Represents all rotation statuses a block can be. This is used both as a rotation state, and to indicate how much relative movement shoud be applied for various functions.