Crate tetrs

The tetrs library provides a number of low-level tasks related to movement of blocks. The code aims to be correct and provide easy extension for new input.

Examples

use tetrs::{block, field, Rotation, Direction};

let field = field::new().width(12);
let block = block::new().on_field(&field);

block.rotate(&field, Rotation::R90);
block.shift_extend(&field, Direction::Down);

Intra-module dependencies are as reduced in scope as possible. For example, a block itself has no knowledge of a wallkick, but provides functionality in terms of rotations with offset to allow for easy use with wallkicks.

A higher-level abstraction in terms of a game-engine is provided. This provides DAS-like behaviour and such.

Modules

block
A single tetrimino.

controller
An abstract controller for specifying actions.

engine
Implements a high-level engine which composes all the components into one abstract structure.

field
A generic playfield.

options
Defines immutable options which are used by an engine.

randomizer
Implements a randomizer.

rotation
This modules provides an interface for dealing with different block offsets.

schema
Methods for converting to and from a textual field representation.

statistics
Stores statistics about the current game.

utility
Contains a number of helper methods which are composed of a number of structures.

wallkick
Implements a wallkick test