Struct tetrs::block::Block

pub struct Block {
    pub x: i32,
    pub y: i32,
    pub id: Type,
    pub r: Rotation,
    pub rs: &'static RotationSystem,
}

A struct representing a single tetrimino.

Blocks are defined by an (x, y) coordinate pair, a Rotation, Type, and an accompanying set of offset data which specifies exactly which pieces in the field this block occupies.

Internally, (x, y) coordinates are taken with their origin from the top right of the grid, to preserve some compatibility with GUI systems. All internal block calculations follow this as well:

(0, 0)-----------------(10, 0)
  |                       |
  |                       |
  |        .(4, 12)       |
  |                       |
  |                       |
  |                       |
(0, 25)----------------(0, 25)

Internally, the (x, y) and data fields look like the following:

  (x, y) = (4, 4)
    data = [(1, 0), (0, 1), (1, 1), (2, 1)]

When calculating a block's position, the data offsets are added to the base coordinates to produce the block. The previous block is a Type::T, with Rotation::R0, as the final data represented is:

 block = [(5, 4), (4, 5), (5, 5), (6, 5)]

It is important to note that offsets are all non-negative, and as such the coordinate effectively partitions the space where the block can reside. Further, a block could have many different internal representations which appear equal, by adjusting the (x, y) coordinates and data in conjunction.

Fields

x

X-coordinate of the piece

y

Y-coordinate of the piece

id

Type of the block

r

Rotation state of the block

rs

Rotation system used internally

Methods

impl Block

fn new(id: Type) -> Block

Construct a new default Block and return it.

The only required value for a block is Type, otherwise the following default values are used:

(x, y)

The default spawn value is (4, 0). This is a common spawn position for fields of standard size.

Rotation

The default rotation is Rotation::R0.

If values are required to be overridden, look at the BlockBuilder trait implementation.

Examples

use tetrs::block::Block;

let block = Block::new(tetrs::block::Type::I);

fn collision(&self, field: &Field) -> bool

Return true if the block currently collides with any pieces on the field.

fn shift(&mut self, field: &Field, direction: Direction) -> bool

Shift the block one step in the specified direction.

Examples

use tetrs::field::Field;
use tetrs::block::{Block, BlockBuilder, Type};
use tetrs::Direction;

let field = Field::new();
let mut block = Block::new(Type::Z)
                      .on_field(&field);
block.shift(&field, Direction::Left);

fn shift_extend(&mut self, field: &Field, direction: Direction)

Repeatedly shift a block as far as we can until a collision occurs. A HardDrop can be performed for example by calling Block.shift_extend(&field, Direction::Down).

Examples

use tetrs::field::Field;
use tetrs::block::{Block, BlockBuilder, Type};

let field = Field::new();
let mut block = Block::new(Type::Z).on_field(&field);
block.shift_extend(&field, tetrs::Direction::Left);

fn rotate_with_offset(&mut self, field: &Field, rotation: Rotation, (x, y): (i32, i32)) -> bool

Rotate the block by a specified amount and then apply an offset.

This is useful for calculating wallkicks, where a collision check is done only after an offset is applied.

use tetrs::field::Field;
use tetrs::block::{Block, BlockBuilder, Type};
use tetrs::Rotation;

let field = Field::new();
let mut block = Block::new(Type::Z).on_field(&field);

// Rotate then move down 2 and right 1 and test for collision
block.rotate_with_offset(&field, Rotation::R90, (2, 1));

fn rotate(&mut self, field: &Field, rotation: Rotation) -> bool

Rotate the block by the specified amount.

Examples

use tetrs::field::Field;
use tetrs::block::{Block, BlockBuilder, Type};
use tetrs::Rotation;

let field = Field::new();
let mut block = Block::new(Type::Z).on_field(&field);
block.rotate(&field, Rotation::R90);

fn at(&self, (a, b): (usize, usize)) -> bool

Check if the block occupies a particular (x, y) absolute location.

fn ghost(&self, field: &Field) -> Block

Return a Block which is a ghost of the current.

Examples

use tetrs::field::Field;
use tetrs::block::{Block, BlockBuilder};

let field = Field::new();
let block = Block::new(tetrs::block::Type::Z).on_field(&field);
let ghost = block.ghost(&field);

fn data(id: Type, r: Rotation) -> &'static [(usize, usize)]

Return the used block data for the specified type.

It may sometimes be useful to query block data without making a block instance itself. For example, when drawing preview pieces from only a Type value.

Examples

use tetrs::block::{Block, Type};
use tetrs::Rotation;

let data = Block::data(Type::Z, Rotation::R270);