Struct tetrs::engine::Engine

pub struct Engine {
    pub controller: Controller,
    pub randomizer: BagRandomizer,
    pub wallkick: &'static WallkickTest,
    pub field: Field,
    pub block: Block,
    pub hold: Option<Block>,
    pub das: u64,
    // some fields omitted
}

This engine allows for handling of DAS-like features and other things which are otherwise transparent to sub-components which are only managed on a per-tick basis (have no concept of state over time).

Fields

controller

Controller which is used by the engine

randomizer

The randomizer being used.

wallkick

The wallkick object being used.

field

The field which the game is played on

block

The active block

hold

The current hold block

das

Das value

Methods

impl Engine

fn new(field: Field, wallkick: &'static WallkickTest) -> Engine

Construct a new engine object and return it.

This should be used as a compositional object of all the underlying objects. This adheres moreso to the rust philosophy and gives greater variance in how an engine can be constructed.

An engine is constructed in an initialized state and is ready to be used right from the beginning.

fn update(&mut self)

The main update phase of the engine.

This handles DAS and all other internal complications based on the current controller state only.

Each call to update is expected to take place in ~mspt ms. It is up to the caller to manage the update lengths appropriately.